pyproject.toml
[project]
name = "yoloviz"
description = "YOLO dataset viewer."
readme = "README.md"
authors = [{ name = "SOUCASSE Bastien", email = "b.soucasse@geo-sat.com" }]
requires-python = ">= 3.12"
dynamic = ["version"]
dependencies = ["pillow", "pyglet", "rich-click"]

[project.scripts]
yoloviz = "yoloviz.cli:cli"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hook.vcs]
version-file = "_version.py"

[tool.hatch.metadata]
allow-direct-dependencies = true

[tool.ruff]
line-length = 120
preview = true

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["ALL"]
ignore = ["CPY001"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
strict = true
